Change some numeric variables from Double to BigDecimal for increased accuracy. Fix calculation of cash remaining.

+++ src/main/scala/com/eigenroute/portfolioanalysis/rebalancing/ETFDesiredValue.scala
@@ -1,3 +1,3 @@
 package com.eigenroute.portfolioanalysis.rebalancing
 
-case class ETFDesiredValue(eTFCode: ETFCode, value: Double, isToTrade: Boolean)
+case class ETFDesiredValue(eTFCode: ETFCode, value: BigDecimal, isToTrade: Boolean)
