+++ src/main/scala/com/eigenroute/portfolioanalysis/rebalancing/PortfolioValueDifference.scala
@@ -1,3 +1,3 @@
 package com.eigenroute.portfolioanalysis.rebalancing
 
-case class PortfolioValueDifference(eTFCode: ETFCode, valueDifference: Double)
+case class PortfolioValueDifference(eTFCode: ETFCode, valueDifference: BigDecimal)
